Delete test/debug.h
