+++ src/common/fr.h
@@ -64,6 +64,7 @@ bool fr_is_null(const fr_t *p);
 void fr_div(fr_t *out, const fr_t *a, const fr_t *b);
 void fr_pow(fr_t *out, const fr_t *a, uint64_t n);
 void fr_from_uint64(fr_t *out, uint64_t n);
+void print_fr(const fr_t *f);
 
 #ifdef __cplusplus
 }
