+++ src/common/g1.h
@@ -44,6 +44,7 @@ extern "C" {
 
 void g1_sub(g1_t *out, const g1_t *a, const g1_t *b);
 void g1_mul(g1_t *out, const g1_t *a, const fr_t *b);
+void print_g1(const g1_t *g);
 
 #ifdef __cplusplus
 }
