+++ src/eip7594/cell.h
@@ -41,3 +41,17 @@
 typedef struct {
     uint8_t bytes[BYTES_PER_CELL];
 } Cell;
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+// Public Functions
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void print_cell(const Cell *cell);
+
+#ifdef __cplusplus
+}
+#endif
