--- src/test/debug.h
@@ -1,40 +1,0 @@
-/*
- * Copyright 2024 Benjamin Edgington
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include "common/bytes.h"
-#include "eip4844/blob.h"
-#include "eip7594/cell.h"
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Public Functions
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void print_bytes32(const Bytes32 *bytes);
-void print_bytes48(const Bytes48 *bytes);
-void print_fr(const fr_t *f);
-void print_g1(const g1_t *g);
-void print_blob(const Blob *blob);
-void print_cell(const Cell *cell);
-
-#ifdef __cplusplus
-}
-#endif
